+++ backend/webtoon-file/src/main/java/com/www/file/service/WebtoonService.java
@@ -6,14 +6,12 @@ import com.www.core.common.Response;
 import com.www.core.file.entity.Episode;
 import com.www.core.file.entity.Webtoon;
 import com.www.core.file.repository.WebtoonRepository;
-import com.www.file.dto.WebtoonDto;
-import com.www.file.dto.WebtoonListDto;
-import com.www.file.dto.WebtoonPage;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.www.file.dto.*;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -26,26 +24,25 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-
 @Service
 public class WebtoonService {
-	
-	private WebtoonRepository webtoonRepository;
-	@Autowired
-	private UserRepository userRepository;
-	
+	private final WebtoonRepository webtoonRepository;
+	private final UserRepository userRepository;
+
+	public WebtoonService(WebtoonRepository webtoonRepository, UserRepository userRepository) {
+		this.webtoonRepository = webtoonRepository;
+		this.userRepository = userRepository;
+	}
+
 	//한 블럭 내 최대 페이지 번호 수
 	private static final int BLOCK_PAGE_NUM_COUNT = 5;
 	//한 페이지 내 최대 웹툰 출력 갯수
 	private static final int PAGE_WEBTOON_COUNT = 10;
+	//한 페이지 내 최대 회차 출력 갯수
+	private static final int PAGE_EPISODE_COUNT = 7;
 	
 	@Value("${custom.path.upload-images}")
 	private String filePath;
-	
-	
-	public WebtoonService(WebtoonRepository webtoonRepository) {
-		this.webtoonRepository = webtoonRepository;
-	}
 	
 	public void checkCondition(MultipartFile file,WebtoonDto webtoonDto,Response<WebtoonDto> res) {
 		
@@ -72,6 +69,66 @@ public class WebtoonService {
 		
 		res.setCode(0);
 		res.setMsg("insert complete");
+	}
+
+	@Transactional
+	public MainWebtoonPage getWebtoonList(Integer pageNum, Response<MainWebtoonPage> res, int sort) {
+
+		Page<Webtoon> page = null;
+		switch(sort) {
+			//기본정렬
+			case 0:
+				page = webtoonRepository.findAll(PageRequest.of(pageNum-1, PAGE_EPISODE_COUNT));
+				break;
+			//조회순 정렬
+			case 1:
+				page = webtoonRepository.findAll(PageRequest.of(pageNum-1, PAGE_EPISODE_COUNT,
+						Sort.by(Sort.Direction.DESC,"hits")));
+				break;
+			//별점 정렬
+			case 2:
+				page = webtoonRepository.findAll(PageRequest.of(pageNum-1, PAGE_EPISODE_COUNT,
+						Sort.by(Sort.Direction.DESC,"epRatingAvg")));
+				break;
+			default:
+				res.setCode(1);
+				res.setMsg("fail : sortNum is not valid");
+				break;
+		}
+
+		List<Webtoon> webtoons = page.getContent();
+		List<MainWebtoonDto> webtoonListDto = new ArrayList<>();
+		int totalpages = page.getTotalPages();
+
+		//등록된 웹툰이 없을 경우
+		if(totalpages == 0 ) totalpages = 1;
+
+		//요청한 페이지 번호가 유효한 범위인지 체크
+		if(pageNum>0 && pageNum<=totalpages) {
+			for(Webtoon webtoon : webtoons) {
+				MainWebtoonDto webtoonDto = MainWebtoonDto.builder()
+						.idx(webtoon.getIdx())
+						.title(webtoon.getTitle())
+						.thumbnail("http://localhost:8081/static/web_thumbnail/"+webtoon.getThumbnail())
+						.storyGenre1(webtoon.getStoryGenre1())
+						.storyGenre2(webtoon.getStoryGenre2())
+						.author(webtoon.getUser().getName())
+						.hits(webtoon.getHits())
+						.epRatingAvg(webtoon.getRatingAvg())
+						.build();
+				webtoonListDto.add(webtoonDto);
+			}
+			res.setCode(0);
+			res.setMsg("show complete");
+		}
+
+		else {
+			res.setCode(1);
+			res.setMsg("fail : pageNum is not in valid range");
+		}
+
+		MainWebtoonPage mainWebtoonPage = new MainWebtoonPage(webtoonListDto, totalpages);
+		return mainWebtoonPage;
 	}
 	
 	@Transactional
